+++ src/Resources/ResourceManager/Implementation/Deployments/TestAzureManagementGroupDeploymentCmdlet.cs
@@ -54,7 +54,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                 QueryString = QueryString,
                 TemplateFile = this.TemplateUri ?? this.TryResolvePath(this.TemplateFile),
                 TemplateObject = this.TemplateObject,
-                TemplateParameterObject = this.GetTemplateParameterObject(this.TemplateParameterObject),
+                TemplateParameterObject = this.GetTemplateParameterObject(),
                 ParameterUri = this.TemplateParameterUri
             };